+++ R/drainageBasins.R
@@ -38,6 +38,7 @@
 #' @param save_path The directory where you want the output shapefiles saved.
 #' @param snap Snap to the "nearest" derived (calculated) stream, or to the "greatest" flow accumulation cell within the snap distance? Beware that "greatest" will move the point downstream by up to the 'snap_dist' specified, while nearest might snap to the wrong stream.
 #' @param snap_dist The search radius within which to snap points to streams. Snapping method depends on 'snap' parameter. Note that distance units will match the projection, so probably best to work on a meter grid.
+#' @param burn_dist If specifying a streams layer polyline layer, the number of DEM units to depress the DEM along the stream trace.
 #' @param n.cores The maximum number of cores to use. Leave NULL to use all cores minus 1.
 #' @param force_update_wbt Whitebox Tools is by default only downloaded if it cannot be found on the computer, and no check are performed to ensure the local version is current. Set to TRUE if you know that there is a new version and you would like to use it.
 #' @param silent_wbt Should Whitebox tools messages be suppressed? This function prints messages to the console already but these messages can be useful if you need to do some debugging.
@@ -67,19 +68,32 @@
 #' }
 
 
-drainageBasins <- function(DEM, streams = NULL, breach_dist = 10000, threshold = 500, overwrite = FALSE, projection = NULL, points = NULL, points_name_col = NULL, save_path = NULL, snap = "nearest", snap_dist = 200, n.cores = NULL, force_update_wbt = FALSE, silent_wbt = TRUE) {
+drainageBasins <- function(DEM, streams = NULL, breach_dist = 10000, threshold = 500, overwrite = FALSE, projection = NULL, points = NULL, points_name_col = NULL, save_path = NULL, snap = "nearest", snap_dist = 200, burn_dist = 10, n.cores = NULL, force_update_wbt = FALSE, silent_wbt = TRUE) {
+
+  # Initial setup #####################################################
+  rlang::check_installed("whitebox", reason = "required to use function drainageBasins") #This is here because whitebox is not a 'depends' of this package; it is only necessary for this function and is therefore in "suggests"
+  invisible(utils::capture.output(wbtCheck(force = force_update_wbt, silent = TRUE)))  #Check whitebox binaries existence and version, install if necessary or if force_update_wbt = TRUE.
 
   if (silent_wbt) {
-    old_option <- options("whitebox.verbose_mode")
-    options("whitebox.verbose_mode" = FALSE)
-    if (old_option$whitebox.verbose_mode) {
-      on.exit(options("whitebox.verbose_mode" = TRUE))
+    old_option <- whitebox::wbt_verbose()
+    whitebox::wbt_verbose(FALSE)
+    if (!is.null(old_option)) {
+      if (old_option) {
+        on.exit(options("whitebox.verbose_mode" = TRUE))
+      }
+    } else {
+      on.exit(whitebox::wbt_verbose(TRUE))
     }
   }
 
-  #initial checks
-  rlang::check_installed("whitebox", reason = "required to use function drainageBasins") #This is here because whitebox is not a 'depends' of this package; it is only necessary for this function and is therefore in "suggests"
-  wbtCheck(force = force_update_wbt, silent = TRUE)  #Check whitebox binaries existence and version, install if necessary or if force_update_wbt = TRUE.
+  # Set warn options to 1 so the user sees warnings as they happen (better context and could enable user to cancell lengthy operation without waiting till the end).
+  old_warn <- options("warn")
+  # See if you can set option without modifying it. If not, don't modify it.
+  try({
+    options(warn = old_warn$warn) #If this succeeds then the option can be modified. If it fails the two lines below are not run and the option is not modified.
+    options(warn = 1)
+    on.exit(options(warn = old_warn$warn))
+  })
 
   if (!(snap %in% c("nearest", "greatest"))) {
     stop("The parameter 'snap' must be one of 'nearest' or 'greatest'.")
@@ -92,7 +106,7 @@ drainageBasins <- function(DEM, streams = NULL, breach_dist = 10000, threshold =
   input_DEM <- DEM #at this point DEM is a path, not an object
   if (!is.null(streams)) {
     if (!file.exists(streams)) {
-      stop("The streams shapefile you pointed to does not exist. Perhaps your file path is wrong? Reminder, I'm looking for the .shp file with extension.")
+      stop("The streams vector file you pointed to does not exist. Perhaps your file path is wrong or is missing the extension?")
     }
   }
 
@@ -117,6 +131,9 @@ drainageBasins <- function(DEM, streams = NULL, breach_dist = 10000, threshold =
       stop("The save_path you provided does not exist. Ceate it and try again.")
     }
     points <- suppressWarnings(terra::vect(points)) #load the points
+    if (!terra::is.points(points)) {
+      stop("The points vector file you specified is not, in fact, of point geometry type.")
+    }
     if (!(points_name_col %in% names(points))) {
       stop("The column name you passed to parameter points_name_col does not appear to be in the points shapefile. If the column name had spaces, slashes, or other problematic characters, it might have been modified upon reading it in. To see what R thinks the column names are you could load the layer using names(terra::vect('path_to_your_shp')).")
     }
@@ -164,7 +181,7 @@ drainageBasins <- function(DEM, streams = NULL, breach_dist = 10000, threshold =
   if (!streams_derived_exists | !d8pntr_exists | !d8fac_exists | overwrite) {
     message("Caculating layers derived from the DEM as they are either missing, have different extents as the provided DEM, you've requested an overwrite of calculated layers, or you specified a streams shapefile.")
 
-    hydroProcess_output <- hydroProcess(DEM = input_DEM, breach_dist = breach_dist, streams = streams, burn_dist = 10, save_path = directory)
+    hydroProcess_output <- hydroProcess(DEM = input_DEM, breach_dist = breach_dist, streams = streams, burn_dist = burn_dist, save_path = directory)
     stream_outputs <- createStreams(DEM = hydroProcess_output, threshold = threshold, save_path = directory)
     d8fac <-  stream_outputs$flow_accum
     d8pntr <- stream_outputs$flow_dir
@@ -205,10 +222,10 @@ drainageBasins <- function(DEM, streams = NULL, breach_dist = 10000, threshold =
     count <- 0 #For 'together' shapefiles. Need a feature to create the R object, then features can be appended.
     failed <- character()
     message(crayon::blue$bold("\n  Starting watershed delineation. This can take a long time so get yourself a tea/coffee.  \n"))
-    for(i in 1:nrow(snapped_points)) {
+    for (i in 1:nrow(snapped_points)) {
       message("Delineating drainage basin for point ", as.data.frame(snapped_points[i, points_name_col]), " (", i, " of ", nrow(snapped_points), ")")
       tryCatch({
-        terra::writeVector(snapped_points[i, ], paste0(tempdir(), "/shapefiles/", i,".shp"), overwrite=TRUE)
+        terra::writeVector(snapped_points[i, ], paste0(tempdir(), "/shapefiles/", i,".shp"), overwrite = TRUE)
         invisible(utils::capture.output(whitebox::wbt_watershed(d8_pntr = paste0(directory, "/D8pointer.tif"),
                                                                 pour_pts = paste0(tempdir(), "/shapefiles/", i, ".shp"),
                                                                 output = paste0(tempdir(), "/rasters/", i, ".tif")
@@ -246,14 +263,14 @@ drainageBasins <- function(DEM, streams = NULL, breach_dist = 10000, threshold =
               nums <- c(nums, as.integer(num))
             }
           }
-          dir.create(paste0(save_path, "/watersheds_", Sys.Date(), "/", folder_name, "_duplicate", if (length(nums) > 0) max(nums)+1 else ""))
+          dir.create(paste0(save_path, "/watersheds_", Sys.Date(), "/", folder_name, "_duplicate", if (length(nums) > 0) max(nums) + 1 else ""))
           save_watershed <<- paste0(save_path, "/watersheds_", Sys.Date(), "/", as.data.frame(snapped_pt[1, points_name_col]), "_duplicate", if (length(nums) > 0) max(nums) + 1 else "")
-          folder_name <<- paste0(folder_name, "_duplicate", if (length(nums) > 0) max(nums)+1 else "")
+          folder_name <<- paste0(folder_name, "_duplicate", if (length(nums) > 0) max(nums) + 1 else "")
         })
 
-        terra::writeVector(snapped_pt, paste0(save_watershed, "/", folder_name, "_snapped_pour_point.shp"), overwrite=TRUE)
-        terra::writeVector(point, paste0(save_watershed, "/", folder_name, "_input_point.shp"), overwrite=TRUE)
-        terra::writeVector(poly, paste0(save_watershed, "/", folder_name, "_drainage_basin.shp"), overwrite=TRUE)
+        terra::writeVector(snapped_pt, paste0(save_watershed, "/", folder_name, "_snapped_pour_point.shp"), overwrite = TRUE)
+        terra::writeVector(point, paste0(save_watershed, "/", folder_name, "_input_point.shp"), overwrite = TRUE)
+        terra::writeVector(poly, paste0(save_watershed, "/", folder_name, "_drainage_basin.shp"), overwrite = TRUE)
 
         #now create/append to the larger shapefiles
         if (count == 0) {
@@ -273,9 +290,9 @@ drainageBasins <- function(DEM, streams = NULL, breach_dist = 10000, threshold =
       })
     }
     #Save the larger shapefiles to disc
-    terra::writeVector(output_basins, paste0(save_path, "/watersheds_", Sys.Date(), "/drainage_basins.shp"), overwrite=TRUE)
-    terra::writeVector(input_points, paste0(save_path, "/watersheds_", Sys.Date(), "/input_points.shp"), overwrite=TRUE)
-    terra::writeVector(snapped_points, paste0(save_path, "/watersheds_", Sys.Date(), "/snapped_points.shp"), overwrite=TRUE)
+    terra::writeVector(output_basins, paste0(save_path, "/watersheds_", Sys.Date(), "/drainage_basins.shp"), overwrite = TRUE)
+    terra::writeVector(input_points, paste0(save_path, "/watersheds_", Sys.Date(), "/input_points.shp"), overwrite = TRUE)
+    terra::writeVector(snapped_points, paste0(save_path, "/watersheds_", Sys.Date(), "/snapped_points.shp"), overwrite = TRUE)
 
     if (length(failed) == nrow(snapped_points)) {
       warning(crayon::red$bold("Failed to delineate all points. Re-check your inputs carefully, and if the issue persists troubleshoot by running the function line by line."))