+++ tests/testthat/test-spatPredict.R
@@ -16,7 +16,7 @@ test_that("single model spatPredict works", {
     allowParallel = TRUE)
 
   res <- spatPredict(features, outcome, 100, trainControl, methods = "ranger", thinFeatures = FALSE, predict = TRUE, n.cores = 2)
-  expect_named(res, c("training_df", "testing_df", "failed_methods", "selected_model", "selected_model_performance", "prediction"))
+  expect_named(res, c("training", "testing", "selected_method", "best_model", "selected_model_performance", "prediction"))
   vdiffr::expect_doppelganger("permafrost prediction single model", terra::plot(res$prediction))
 })
 
@@ -44,6 +44,6 @@ test_that("multi-model spatPredict works", {
   )
 
   res <- suppressWarnings(spatPredict(features, outcome, 100, trainControl, methods = c("ranger", "Rborist"), thinFeatures = FALSE, predict = TRUE, n.cores = 2))
-  expect_named(res, c("training_df", "testing_df", "failed_methods", "trained_models_performance", "selected_model", "selected_model_performance", "prediction"))
+  expect_named(res, c("training", "testing", "failed_methods", "warned_methods", "warned_methods", "error_messages", "warn_messages", "trained_models_performance", "best_model", "selected_model_performance", "prediction"))
   vdiffr::expect_doppelganger("permafrost prediction two models", terra::plot(res$prediction))
 })